# Optimise Iterator::{max, max_by, min, min_by}.